backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain-google-genai
langchain
faiss-cpu
langchain-community
tiktoken
pydantic
